requirements.txt
mysql-connector-python           8.3.0
Flask                            2.2.2
Werkzeug                         2.2.3